Refactor search logic for re-use
