+++ WTG.Analyzers/Analyzers/BooleanLiteral/BooleanLiteralAnalyzer.cs
@@ -1,11 +1,9 @@
-using System;
 using System.Collections.Immutable;
-using System.Threading;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.Text;
+using WTG.Analyzers.Analyzers.BooleanLiteral;
 using WTG.Analyzers.Utils;
 
 namespace WTG.Analyzers
@@ -51,23 +49,9 @@ namespace WTG.Analyzers
 				return;
 			}
 
-			var methodInvocation = argument.FirstAncestorOrSelf<InvocationExpressionSyntax>();
-			if (methodInvocation is null)
-			{
-				return;
-			}
+			var argumentSymbol = argument.TryFindCorrespondingParameterSymbol(context.SemanticModel, context.CancellationToken);
 
-			var methodSymbol = (IMethodSymbol)context.SemanticModel.GetSymbolInfo(methodInvocation, context.CancellationToken).Symbol;
-			if (methodSymbol is null)
-			{
-				return;
-			}
-
-			var argumentList = (ArgumentListSyntax)argument.Parent;
-			var index = FindArgumentIndex(argumentList, argument);
-			var argumentSymbol = methodSymbol.Parameters[index];
-
-			if (argumentSymbol.OriginalDefinition.Type.TypeKind == TypeKind.TypeParameter)
+			if (argumentSymbol is null || argumentSymbol.OriginalDefinition.Type.TypeKind == TypeKind.TypeParameter)
 			{
 				return;
 			}
@@ -76,23 +60,6 @@ namespace WTG.Analyzers
 				Diagnostic.Create(
 					Rules.UseNamedArgumentsWhenPassingBooleanLiteralsRule,
 					argument.GetLocation()));
-		}
-
-		static int FindArgumentIndex(ArgumentListSyntax haystack, ArgumentSyntax needle)
-		{
-			var i = 0;
-			var enumerator = haystack.Arguments.GetEnumerator();
-			while (enumerator.MoveNext())
-			{
-				if (enumerator.Current == needle)
-				{
-					return i;
-				}
-
-				i++;
-			}
-
-			throw new InvalidOperationException("Failed to find Argument in its parent ArgumentList.");
 		}
 	}
 }
